refactor: restructure schema & entity types
