+++ src/lib/external/mud/entity.ts
@@ -1,6 +1,4 @@
-import { concatHex, decodeAbiParameters, encodeAbiParameters, type Hex, isHex, size, sliceHex } from "viem";
-
-import type { AbiToSchema, AbiKeySchema, Properties, SchemaToPrimitives } from "@/lib/external/mud/schema";
+import { concatHex, type Hex, isHex, size, sliceHex } from "viem";
 
 /**
  * An entity is a string that represents a tuple of hex keys, to identify a row in a table.
@@ -51,57 +49,6 @@ export function entityToHexKeyTuple(entity: Entity): readonly Hex[] {
   }
   return new Array(length / 32).fill(0).map((_, index) => sliceHex(entity, index * 32, (index + 1) * 32));
 }
-
-// Valid keys to use as schema
-
-/**
- * Concatenate a tuple of hex keys into a single entity, after encoding them into hex strings.
- *
- * Note: This is especially useful when trying to retrieve an entity using its separate key properties, as it is done in the table key
- * methods attached to contract tables (see {@link createTableKeyMethods}).
- *
- * @category Entity
- */
-export const encodeEntity = <TKeySchema extends AbiKeySchema, T = unknown>(
-  abiKeySchema: TKeySchema,
-  keys: Properties<AbiToSchema<TKeySchema>, T>,
-) => {
-  if (Object.keys(abiKeySchema).length !== Object.keys(keys).length) {
-    throw new Error(
-      `entity length ${Object.keys(keys).length} does not match entity schema length ${Object.keys(abiKeySchema).length}`,
-    );
-  }
-
-  return hexKeyTupleToEntity(
-    Object.entries(abiKeySchema).map(([keyName, type]) => encodeAbiParameters([{ type }], [keys[keyName]])),
-  );
-};
-
-/**
- * Decode an entity into a tuple of hex keys, after decoding them from hex strings.
- *
- * Note: This is useful for retrieving the values of each separate key property from an entity, using its schema and actual entity string.
- *
- * @category Entity
- */
-export const decodeEntity = <TKeySchema extends AbiKeySchema>(
-  abiKeySchema: TKeySchema,
-  entity: Entity,
-): SchemaToPrimitives<TKeySchema> => {
-  const hexKeyTuple = entityToHexKeyTuple(entity);
-  if (hexKeyTuple.length !== Object.keys(abiKeySchema).length) {
-    throw new Error(
-      `entity entity tuple length ${hexKeyTuple.length} does not match entity schema length ${Object.keys(abiKeySchema).length}`,
-    );
-  }
-
-  return Object.fromEntries(
-    Object.entries(abiKeySchema).map(([entity, type], index) => [
-      entity,
-      decodeAbiParameters([{ type }], hexKeyTuple[index] as Hex)[0],
-    ]),
-  ) as SchemaToPrimitives<TKeySchema>;
-};
 
 /**
  * Get the symbol corresponding to an entity's hex ID.
