+++ src/lib/external/mud/tables.ts
@@ -1,14 +1,10 @@
 import { map, pipe } from "rxjs";
 import isEqual from "fast-deep-equal";
 
-import type { IndexedBaseTable, BaseTable } from "@/tables/types";
+import type { IndexedBaseTable, BaseTable, TableMutationOptions } from "@/tables/types";
 import type { TableUpdate } from "@/queries/types";
-import { getEntitySymbol, type Entity } from "@/lib/external/mud/entity";
+import { type Entity, getEntitySymbol } from "@/lib/external/mud/entity";
 import { type BaseTableMetadata, OptionalTypes, type Properties, type Schema } from "@/lib/external/mud/schema";
-
-export type TableMutationOptions = {
-  skipUpdateStream?: boolean;
-};
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 function getTableName(table: BaseTable<any, any, any>) {
